DDS TypeObject construction in a separate file
