+++ rmw_cyclonedds_cpp/src/rmw_node.cpp
@@ -99,6 +99,7 @@
 #include "serdata.hpp"
 #include "demangle.hpp"
 
+#include "dyntype.hpp"
 
 using namespace std::literals::chrono_literals;
